cleanup: Moving components to components folder for better project structure
